// src/api/generated/model/contestTeamMemberQuestionAnalytics.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionDifficulty } from './questionDifficulty';

export interface ContestTeamMemberQuestionAnalytics {
  question_id: string;
  title: string;
  difficulty: QuestionDifficulty;
  time_limit_ms: number;
  memory_limit_mb: number;
  total_submission?: number;
  accepted_submission?: number;
}