// src/api/generated/model/contestTeamMemberQuestionSubmissionItem.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { SubmissionStatus } from './submissionStatus';

export interface ContestTeamMemberQuestionSubmissionItem {
  submission_id: string;
  status?: SubmissionStatus | null;
  score?: number;
  language: string;
  created_at: string;
  execution_time?: number | null;
  memory?: number | null;
}