// src/api/generated/model/submissionTestCaseResultSchema.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */
import type { SubmissionStatus } from './submissionStatus';

/**
 * Single testcase result for a submission.
 */
export interface SubmissionTestCaseResultSchema {
  id: string;
  name: string;
  status?: SubmissionStatus | null;
  execution_time?: number | null;
  memory?: number | null;
  input: string;
  expected_output: string;
  actual_output?: string | null;
}